Prompt user on db upgrade failure
Previously, we dropped and recreated database, which is bad for users
who want to keep their content and didn't cause this situation on
purpose.

So instead, we show a dialog to the user explaining the situation, and
prompting to an app upgrade or to manually clear app data.

This also allows us to not care about the removal or backup of
attachments in the case of upgrade failure.

Change-Id: I64e0e6b91272ba750f7f1c6c2df024e970419d4b